new map design, and added stuff to thinking
i made it so when you view the map you see the first letter of the area in order to make it easier to understand what is next to the player. (just an idea we can change it back to keep a mystery.  we could have it so when the player  visits an area the area letter is revealed)